backend/requirements.txt
Django>=5.0,<5.1
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.3.0
psycopg2-binary>=2.9.9
django-cors-headers>=4.3.1
Pillow>=10.1.0
paypalrestsdk>=1.13.1
python-decouple>=3.8
